add cargo lock for reproducible builds
For reproducible builds we need to guarantee to build the final product
using the same input dependencies. This feature is supported by cargo
using Cargo.lock and the according --locked CLI options.

This requires to commit the lock file whenever updates are managed or
modified so we have a frozen set.

https://doc.rust-lang.org/cargo/guide/cargo-toml-vs-cargo-lock.html
https://doc.rust-lang.org/cargo/faq.html#why-do-binaries-have-cargolock-in-version-control-but-not-libraries

Why does it matter?
https://reproducible-builds.org/